WIP: move files; scripts are working
