+++ .travis.yml
@@ -2,6 +2,5 @@ language: node_js
 node_js:
   - '20'
   - '18'
-  - '16'
 os:
   - linux
