+++ packages/countries/src/countries.emoji.data.ts
@@ -1,4 +1,4 @@
 import { TCountryToString } from './types.ts'
-import countriesEmoji from '../dist/more/countries.emoji.min.json'
+import countriesEmoji from '../../../dist/more/countries.emoji.min.json'
 
 export default countriesEmoji as TCountryToString
