+++ packages/countries/src/data/countries.2to3.ts
@@ -1,4 +1,4 @@
-import { TCountryCode } from 'src/types.ts'
+import { TCountryCode } from '../types.ts'
 
 export default {
   AD: 'AND',
