+++ packages/countries/src/index.ts
@@ -1,5 +1,5 @@
-export { default as continents } from 'src/data/continents.ts'
-export { default as countries } from 'src/data/countries.ts'
-export { default as languages } from 'src/data/languages.ts'
+export { default as continents } from './data/continents.ts'
+export { default as countries } from './data/countries.ts'
+export { default as languages } from './data/languages.ts'
 
-export { getEmojiFlag } from 'src/getEmojiFlag.ts'
+export { getEmojiFlag } from './getEmojiFlag.ts'
