+++ packages/countries/tsconfig.json
@@ -1,10 +1,3 @@
 {
-  "extends": "tsconfig/tsconfig.json",
-  "compilerOptions": {
-    "paths": {
-      "src/*": ["./src/*"]
-    }
-  },
-  "include": ["src"],
-  "exclude": ["node_modules"]
+  "extends": "tsconfig/tsconfig.json"
 }
