+++ packages/scripts/package.json
@@ -11,6 +11,7 @@
     "@types/node": "20.0.0",
     "@types/semver-compare": "1.0.1",
     "chalk": "5.2.0",
+    "countries-list": "*",
     "pretty-bytes": "6.1.0",
     "semver-compare": "1.0.0",
     "tsx": "3.12.7"
@@ -19,7 +20,7 @@
     "node": ">=14.16"
   },
   "scripts": {
-    "build-data": "tsx ./index.ts",
-    "version": "tsx ./version.ts"
+    "build-data": "tsx index.ts",
+    "version": "tsx version.ts"
   }
 }
